+++ changelog.md
@@ -2,6 +2,7 @@
 - Remove dependency of `mplutils`
 - Add `Hist1d/2d.copy`
 - Add `Hist2d.from_image`
+- Add tutorial
 
 # v5.4.1
 - Expose `for_pcolormesh_from_txt` and `for_pcolormesh_from_root`